requirements.txt
selenium
python-dotenv
urllib3